+++ commercetools-models/src/test/java/io/sphere/sdk/projects/commands/ProjectCartValueIntegrationTest.java
@@ -1,0 +1,23 @@
+package io.sphere.sdk.projects.commands;
+
+import io.sphere.sdk.projects.CartValue;
+import io.sphere.sdk.projects.Project;
+import io.sphere.sdk.projects.commands.updateactions.SetShippingRateInputType;
+import io.sphere.sdk.projects.queries.ProjectGet;
+import org.junit.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ProjectCartValueIntegrationTest extends ProjectIntegrationTest {
+
+    @Test
+    public void execution() {
+
+        final Project project = client().executeBlocking(ProjectGet.of());
+        final Project updatedProjectCartValue = client().executeBlocking(ProjectUpdateCommand.of(project, SetShippingRateInputType.of(CartValue.of())));
+        assertThat(updatedProjectCartValue.getShippingRateInputType()).isNotNull();
+        assertThat(updatedProjectCartValue.getShippingRateInputType().getType()).isEqualTo(CartValue.of().getType());
+
+    }
+
+}
